Creative: Group Events Complete

+++ delete.php
@@ -13,6 +13,32 @@
     $data = json_decode(file_get_contents('php://input'), true);
     if(hash_equals($_SESSION['token'], $data['token'])){
         $event_id = htmlentities($data['event_id']);
+        
+        // check if the logged in user is the event's creator --> allow deletion, otherwise exit (they shouldn't make it this far, but ensure NO deletion if they do)
+        $stmt = $mysqli->prepare("SELECT creator FROM events WHERE event_id = ?");
+        if(!$stmt){
+            echo json_encode(array(
+                "success" => false,
+                "error" => "Query Preparation Failed: " . $mysqli->error
+            ));
+            exit();
+        }
+
+        $stmt->bind_param('i', $event_id);
+        $stmt->execute();
+        $stmt->bind_result($originalCreator);
+        $stmt->fetch();
+        $stmt->close();
+
+        if($username !== $originalCreator){
+            echo json_encode(array(
+                "success" => false,
+                "error" => "Not Original Event Creator"
+            ));
+            exit();
+        }
+
+        // logged in user is original event creator --> carry on deleting
 
         // grab count of events
         $stmt = $mysqli->prepare("DELETE FROM events WHERE creator = ? AND event_id = ?");
